+++ pkg/bitesize/environment_test.go
@@ -6,6 +6,7 @@ import (
 	"testing"
 
 	"github.com/pearsontechnology/environment-operator/pkg/config"
+	"github.com/stretchr/testify/assert"
 )
 
 func TestExistingEnvironment(t *testing.T) {
@@ -13,7 +14,7 @@ func TestExistingEnvironment(t *testing.T) {
 	e, err := LoadEnvironment("../../test/assets/environments.bitesize", "environment2")
 
 	if err != nil {
-		t.Errorf("Unexpected error loading environment: %s", err.Error())
+		t.Errorf("TestExistingEnvironment: Unexpected error loading environment: %s", err.Error())
 	}
 
 	if len(e.Services) != 7 {
@@ -79,10 +80,88 @@ func TestEnvironmentImportConfigMap(t *testing.T) {
 
 	// fails on travis because .git in travis
 	if err != nil {
-		t.Errorf("Unexpected error loading environment: %s", err.Error())
+		t.Errorf("TestEnvironmentImportConfigMap: Unexpected error loading environment: %s", err.Error())
 	}
 
 	if len(e.Gists) != 3 {
 		t.Errorf("Unexpected count of import. Expected 3, got: %d", len(e.Gists))
 	}
+}
+
+func TestFindVolByName(t *testing.T) {
+	volumes := []Volume{
+		{
+			Name: "zzz",
+			Type: "secret",
+			Path: "path-zzz",
+		},
+		{
+			Name:  "aaa",
+			Type:  "efs",
+			Path:  "path-aaa",
+			Size:  "10G",
+			Modes: "ReadWriteMany",
+		},
+		{
+			Name:  "mmm",
+			Size:  "10G",
+			Path:  "path-mmm",
+			Modes: "ReadWriteOnce",
+		},
+	}
+
+	vol, err := findVolByName(volumes, "aaa")
+	if err != nil {
+		t.Errorf("Could not find volume name: %s", "aaa")
+	}
+	assert.Equal(t, vol, volumes[1], "The returned volume should be the same")
+}
+
+func TestSortVolumesByVolName(t *testing.T) {
+	volumesInput := []Volume{
+		{
+			Name: "zzz",
+			Type: "secret",
+			Path: "path-zzz",
+		},
+		{
+			Name:  "aaa",
+			Type:  "efs",
+			Path:  "path-aaa",
+			Size:  "10G",
+			Modes: "ReadWriteMany",
+		},
+		{
+			Name:  "mmm",
+			Size:  "10G",
+			Path:  "path-mmm",
+			Modes: "ReadWriteOnce",
+		},
+	}
+
+	volumesExpected := []Volume{
+		{
+			Name:  "aaa",
+			Type:  "efs",
+			Path:  "path-aaa",
+			Size:  "10G",
+			Modes: "ReadWriteMany",
+		},
+		{
+			Name:  "mmm",
+			Size:  "10G",
+			Path:  "path-mmm",
+			Modes: "ReadWriteOnce",
+		},
+		{
+			Name: "zzz",
+			Type: "secret",
+			Path: "path-zzz",
+		},
+	}
+	volumesSorted, err := SortVolumesByVolName(volumesInput)
+	if err != nil {
+		t.Errorf("Could not sort volumes: %s", err)
+	}
+	assert.Equal(t, volumesExpected, volumesSorted, "The volumes should be sorted")
 }